+++ src/game/tests.rs
@@ -321,6 +321,7 @@ mod vintage_cube_ajani;
 mod vintage_cube_amped_raptor;
 mod vintage_cube_animate_dead;
 mod vintage_cube_archon_of_cruelty;
+mod vintage_cube_arena_of_glory;
 mod vintage_cube_arrival;
 mod vintage_cube_artifacts;
 mod vintage_cube_atraxa_grand_unifier;